Additional tests for minimal measures
