+++ cqf-fhir-cr/src/test/resources/org/opencds/cqf/fhir/cr/measure/dstu3/MinimalMeasure/cql/Minimal.cql
@@ -1,6 +1,1 @@
 library Minimal
-
-parameter "Measurement Period" Interval<DateTime>
-  default Interval[@2019-01-01T00:00:00.0, @2020-01-01T00:00:00.0)
-
-define "Initial Population": true
