+++ cqf-fhir-cr/src/test/resources/org/opencds/cqf/fhir/cr/measure/dstu3/MinimalMeasure/cql/MinimalWithParameter.cql
@@ -1,0 +1,5 @@
+library MinimalWithParameter
+
+parameter "Measurement Period" Interval<DateTime>
+
+define "Initial Population": true
