+++ cqf-fhir-cr/src/test/resources/org/opencds/cqf/fhir/cr/measure/dstu3/MinimalMeasure/cql/MinimalWithPopulation.cql
@@ -1,0 +1,3 @@
+library MinimalWithPopulation
+
+define "Initial Population": true
