+++ cqf-fhir-cr/src/test/resources/org/opencds/cqf/fhir/cr/measure/dstu3/MinimalMeasure/resources/measure/MinimalWithGroup.json
@@ -1,0 +1,20 @@
+{
+    "id": "MinimalWithGroup",
+    "resourceType": "Measure",
+    "library": [
+        {
+            "reference": "Library/MinimalWithGroup"
+        }
+    ],
+    "scoring": {
+        "coding": [
+            {
+                "system": "http://hl7.org/fhir/measure-scoring",
+                "code": "proportion"
+            }
+        ]
+    },
+    "group": [
+        {}
+    ]
+}
