+++ cqf-fhir-cr/src/test/resources/org/opencds/cqf/fhir/cr/measure/dstu3/MinimalMeasure/resources/measure/MinimalWithParameter.json
@@ -1,0 +1,37 @@
+{
+    "id": "MinimalWithParameter",
+    "resourceType": "Measure",
+    "library": [{
+        "reference" : "Library/MinimalWithParameter"
+    }],
+    "scoring": {
+        "coding": [
+            {
+                "system": "http://hl7.org/fhir/measure-scoring",
+                "code": "proportion"
+            }
+        ]
+    },
+    "group": [
+        {
+            "population": [
+                {
+                    "id": "initial-population",
+                    "code": {
+                        "coding": [
+                            {
+                                "system": "http://terminology.hl7.org/CodeSystem/measure-population",
+                                "code": "initial-population",
+                                "display": "Initial Population"
+                            }
+                        ]
+                    },
+                    "criteria": {
+                        "language": "text/cql-identifier",
+                        "expression": "Initial Population"
+                    }
+                }
+            ]
+        }
+    ]
+}
