+++ cqf-fhir-cr/src/test/resources/org/opencds/cqf/fhir/cr/measure/r4/MinimalMeasure/cql/MinimalWithGroup.cql
@@ -1,0 +1,1 @@
+library MinimalWithGroup
