+++ cqf-fhir-cr/src/test/resources/org/opencds/cqf/fhir/cr/measure/r4/MinimalMeasure/cql/MinimalWithPopulation.cql
@@ -1,0 +1,3 @@
+library MinimalWithPopulation
+
+define "Initial Population": true
