+++ cqf-fhir-cr/src/test/resources/org/opencds/cqf/fhir/cr/measure/r4/MinimalMeasure/resources/library/MinimalWithParameterjson
@@ -1,0 +1,17 @@
+{
+    "resourceType": "Library",
+    "id": "MinimalWithParameter",
+    "url": "http://example.com/Library/MinimalWithParameter",
+    "name": "MinimalWithParameter",
+    "status": "active",
+    "type": {
+      "coding": [ {
+        "system": "http://terminology.hl7.org/CodeSystem/library-type",
+        "code": "logic-library"
+      } ]
+    },
+    "content": [ {
+      "contentType": "text/cql",
+      "url": "../../cql/MinimalWithParameter.cql"
+    } ]
+}
