+++ cqf-fhir-cr/src/test/resources/org/opencds/cqf/fhir/cr/measure/r4/MinimalMeasure/resources/library/MinimalWithPopulation.json
@@ -1,0 +1,17 @@
+{
+    "resourceType": "Library",
+    "id": "MinimalWithPopulation",
+    "url": "http://example.com/Library/MinimalWithPopulation",
+    "name": "MinimalWithPopulation",
+    "status": "active",
+    "type": {
+      "coding": [ {
+        "system": "http://terminology.hl7.org/CodeSystem/library-type",
+        "code": "logic-library"
+      } ]
+    },
+    "content": [ {
+      "contentType": "text/cql",
+      "url": "../../cql/MinimalWithPopulation.cql"
+    } ]
+}
